Refactor: 모델 분리
